+++ common/SetupMessage.hpp
@@ -13,7 +13,7 @@ public:
 
 	typedef std::unordered_map<std::string, int> DataMap;
 
-	SetupMessage(uint16_t id, GameType gType, int pCount, int time, int score, DataMap&& data = DataMap());
+	SetupMessage(uint16_t id, GameType gType, int8_t pCount, int16_t time, int16_t score, DataMap&& data = DataMap());
 
 #ifdef WITH_JSON
 	static std::unique_ptr<SetupMessage> fromJSON(const Json::Value& object);
@@ -27,15 +27,15 @@ public:
 	const GameType gameType;
 
 	/// The number of players playing.
-	const int playerCount;
+	const int8_t playerCount;
 
 	/// If time is a win condition, this is the maximum game length, in seconds.
 	/// If not, -1.
-	const int gameLength;
+	const int16_t gameLength;
 
 	/// If points is a win condition, this is the score a player must get to win.
 	/// If not, -1
-	const int winningScore;
+	const int16_t winningScore;
 
 	const DataMap gameData;
