+++ common/StatusResponseMessage.cpp
@@ -48,7 +48,7 @@ StatusResponseMessage::PlayerList parseStats(const Value& stats)
 } // End anonymous namespace
 
 StatusResponseMessage::StatusResponseMessage(uint16_t id, uint16_t respTo, const std::string& message,
-	                                         bool isRunning, int timeLeft, int winScore, PlayerList&& playerStats) :
+	                                         bool isRunning, int16_t timeLeft, int16_t winScore, PlayerList&& playerStats) :
 	// If we're sending a full status response message back, the request was ok.
 	ResponseMessage(id, respTo, ResponseMessage::Code::OK, message),
 	running(isRunning),
@@ -88,7 +88,8 @@ std::unique_ptr<StatusResponseMessage> StatusResponseMessage::fromJSON(const Jso
 
 	return std::unique_ptr<StatusResponseMessage>(
 		new StatusResponseMessage(responseInfo->id, responseInfo->respondingTo, responseInfo->message,
-		                          runningValue.asBool(), timeRemainingValue.asInt(), winningScoreValue.asInt(),
+		                          runningValue.asBool(),
+		                          (int16_t)timeRemainingValue.asInt(), (int16_t)winningScoreValue.asInt(),
 		                          parseStats(playerStatsValue)));
 }
