+++ common/StatusResponseMessage.hpp
@@ -10,10 +10,10 @@ class StatusResponseMessage : public ResponseMessage {
 public:
 
 	struct PlayerStats {
-		const int score;
-		const int hits;
+		const int16_t score;
+		const int16_t hits;
 
-		PlayerStats(int s, int h) : score(s), hits(h) { }
+		PlayerStats(int16_t s, int16_t h) : score(s), hits(h) { }
 
 		bool operator==(const PlayerStats& o) const { return score == o.score && hits == o.hits; }
 	};
@@ -21,7 +21,7 @@ public:
 	typedef std::vector<PlayerStats> PlayerList;
 
 	StatusResponseMessage(uint16_t id, uint16_t respTo, const std::string& message,
-	                      bool isRunning, int timeLeft, int winScore, PlayerList&& playerStats);
+	                      bool isRunning, int16_t timeLeft, int16_t winScore, PlayerList&& playerStats);
 
 #ifdef WITH_JSON
 	static std::unique_ptr<StatusResponseMessage> fromJSON(const Json::Value& object);
@@ -33,9 +33,9 @@ public:
 
 	const bool running;
 
-	const int timeRemaining;
+	const int16_t timeRemaining;
 
-	const int winningScore;
+	const int16_t winningScore;
 
 	const PlayerList players;
